+++ lib/src/requests/classify_document_request.dart
@@ -1,7 +1,7 @@
 /*
  * --------------------------------------------------------------------------------
  * <copyright company="Aspose" file="classify_document_request.dart">
- *   Copyright (c) 2021 Aspose.Words for Cloud
+ *   Copyright (c) 2022 Aspose.Words for Cloud
  * </copyright>
  * <summary>
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -48,8 +48,11 @@ class ClassifyDocumentRequest implements RequestBase {
   /// Encoding that will be used to load an HTML (or TXT) document if the encoding is not specified in HTML.
   final String loadEncoding;
 
-  /// Password for opening an encrypted document.
+  /// Password of protected Word document. Use the parameter to pass a password via SDK. SDK encrypts it automatically. We don't recommend to use the parameter to pass a plain password for direct call of API.
   final String password;
+
+  /// Password of protected Word document. Use the parameter to pass an encrypted password for direct calls of API. See SDK code for encyption details.
+  final String encryptedPassword;
 
   /// The number of the best classes to return.
   final String bestClassesCount;
@@ -57,7 +60,7 @@ class ClassifyDocumentRequest implements RequestBase {
   /// The taxonomy to use.
   final String taxonomy;
 
-  ClassifyDocumentRequest(final this.name, {final this.folder, final this.storage, final this.loadEncoding, final this.password, final this.bestClassesCount, final this.taxonomy});
+  ClassifyDocumentRequest(final this.name, {final this.folder, final this.storage, final this.loadEncoding, final this.password, final this.encryptedPassword, final this.bestClassesCount, final this.taxonomy});
 
   @override
   Future<ApiRequestData> createRequestData(final ApiClient _apiClient) async {
@@ -83,6 +86,10 @@ class ClassifyDocumentRequest implements RequestBase {
 
     if (password != null) {
       _queryParams['encryptedPassword'] = await _apiClient.encryptPassword(password);
+    }
+
+    if (encryptedPassword != null) {
+      _queryParams['encryptedPassword'] = _apiClient.serializeToString(encryptedPassword);
     }
 
     if (bestClassesCount != null) {