+++ lib/src/requests/get_document_statistics_online_request.dart
@@ -1,7 +1,7 @@
 /*
  * --------------------------------------------------------------------------------
  * <copyright company="Aspose" file="get_document_statistics_online_request.dart">
- *   Copyright (c) 2021 Aspose.Words for Cloud
+ *   Copyright (c) 2022 Aspose.Words for Cloud
  * </copyright>
  * <summary>
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -42,8 +42,11 @@ class GetDocumentStatisticsOnlineRequest implements RequestBase {
   /// Encoding that will be used to load an HTML (or TXT) document if the encoding is not specified in HTML.
   final String loadEncoding;
 
-  /// Password for opening an encrypted document.
+  /// Password of protected Word document. Use the parameter to pass a password via SDK. SDK encrypts it automatically. We don't recommend to use the parameter to pass a plain password for direct call of API.
   final String password;
+
+  /// Password of protected Word document. Use the parameter to pass an encrypted password for direct calls of API. See SDK code for encyption details.
+  final String encryptedPassword;
 
   /// The flag indicating whether to include comments from the WordCount. The default value is "false".
   final bool includeComments;
@@ -54,7 +57,7 @@ class GetDocumentStatisticsOnlineRequest implements RequestBase {
   /// The flag indicating whether to include shape's text from the WordCount. The default value is "false".
   final bool includeTextInShapes;
 
-  GetDocumentStatisticsOnlineRequest(final this.document, {final this.loadEncoding, final this.password, final this.includeComments, final this.includeFootnotes, final this.includeTextInShapes});
+  GetDocumentStatisticsOnlineRequest(final this.document, {final this.loadEncoding, final this.password, final this.encryptedPassword, final this.includeComments, final this.includeFootnotes, final this.includeTextInShapes});
 
   @override
   Future<ApiRequestData> createRequestData(final ApiClient _apiClient) async {
@@ -68,6 +71,10 @@ class GetDocumentStatisticsOnlineRequest implements RequestBase {
 
     if (password != null) {
       _queryParams['encryptedPassword'] = await _apiClient.encryptPassword(password);
+    }
+
+    if (encryptedPassword != null) {
+      _queryParams['encryptedPassword'] = _apiClient.serializeToString(encryptedPassword);
     }
 
     if (includeComments != null) {